Updated ‘_plot’ method in AngleConstraints, ImproperAngleConstraints, BondConstraints classes to enhance tick handling and plotting logic. - Added logic to ensure xticks and xtick labels match in length - Updated tick handling to use FixedLocator and FixedFormatter for better control - Improved histogram rescaling and stacking logic for horizontal stacking - Enhanced tick color adaptation for color-coded x-tick labels - Updated axis label and title setting with additional parameters - Added detailed comments to explain changes and new logic.
Updated ‘__get_total_Gr’ in PairDistributionConstraints class to check if the shape function dimension ‘self._shapeArray’ matches the fit data range for broadcasting. If the fit range changes during the steps of the fit run, these we a broadcasting error that kills the program.  Fixed this to recompute the shape function before applying to obtain total G(r).

Update Constraints method to use ‘float’ and ‘int’ in place of np.int and np.float, as these are deprecated.